requirements.txt
Flask==2.0.1
tensorflow-cpu==2.15.0
numpy==1.24.3
Pillow==10.0.0
gunicorn==21.2.0
Werkzeug==2.0.1
python-dotenv==1.0.0 